+++ android/src/main/kotlin/com/victorcarreras/pdf_combiner/subclasses/CreatePDFFromMultipleImage.kt
@@ -8,64 +8,76 @@ import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
-import java.io.ByteArrayOutputStream
 import java.io.File
 import java.io.FileOutputStream
 import java.io.IOException
 import kotlin.math.roundToInt
 
+class ImageScale(
+    val maxWidth: Int,
+    val maxHeight: Int,
+)
+
+class CompressionLevel(val value: Int)
+class PdfFromMultipleImageConfig(val rescale: ImageScale, val keepAspectRatio: Boolean)
+
+
 class CreatePDFFromMultipleImage(getResult: MethodChannel.Result) {
 
     private var result: MethodChannel.Result = getResult
 
-
     @OptIn(DelicateCoroutinesApi::class)
     fun create(
-        paths: List<String>,
-        outputDirPath: String,
-        needImageCompressor: Boolean,
-        maxWidth: Int,
-        maxHeight: Int
+        inputPaths: List<String>,
+        outputPath: String,
+        config: PdfFromMultipleImageConfig,
     ) {
         var status = ""
 
         val pdfFromMultipleImage = GlobalScope.launch(Dispatchers.IO) {
             try {
-                val file = File(outputDirPath)
+                val file = File(outputPath)
                 val fileOutputStream = FileOutputStream(file)
                 val pdfDocument = PdfDocument()
                 val i = 0
-                for (item in paths) {
+                for (item in inputPaths) {
 
-                    var bitmap = if (needImageCompressor) {
-                        compressImage(item, maxWidth, maxHeight)
+                    val bitmap = compressImage(
+                        item, config.rescale.maxWidth, config.rescale.maxHeight,
+                        config.keepAspectRatio
+                    )
+                    if (bitmap != null) {
+                        val scaledBitmap = scaleBitmap(bitmap, config.rescale.maxWidth)
+                        val pageInfo =
+                            PdfDocument.PageInfo.Builder(
+                                scaledBitmap.width,
+                                scaledBitmap.height,
+                                i + 1
+                            ).create()
+                        val page = pdfDocument.startPage(pageInfo)
+                        val canvas = page.canvas
+                        val paint = Paint()
+                        canvas.drawPaint(paint)
+                        canvas.drawBitmap(scaledBitmap, 0f, 0f, paint)
+                        pdfDocument.finishPage(page)
+                        scaledBitmap.recycle()
+                        pdfDocument.writeTo(fileOutputStream)
+                        status = "success"
                     } else {
-                        BitmapFactory.decodeFile(item)
+                        status = "error"
                     }
 
-                    val pageInfo =
-                        PdfDocument.PageInfo.Builder(bitmap!!.width, bitmap.height, i + 1).create()
-                    val page = pdfDocument.startPage(pageInfo)
-                    val canvas = page.canvas
-                    val paint = Paint()
-                    canvas.drawPaint(paint)
-                    canvas.drawBitmap(bitmap, 0f, 0f, paint)
-                    pdfDocument.finishPage(page)
-                    bitmap.recycle()
                 }
-                pdfDocument.writeTo(fileOutputStream)
                 pdfDocument.close()
-                status = "success"
             } catch (e: IOException) {
                 e.printStackTrace()
                 status = "error"
-
             }
         }
 
         pdfFromMultipleImage.invokeOnCompletion {
             if (status == "success")
-                status = outputDirPath
+                status = outputPath
             else if (status == "error")
                 status = "error"
 
@@ -75,7 +87,19 @@ class CreatePDFFromMultipleImage(getResult: MethodChannel.Result) {
         }
     }
 
-    private fun compressImage(imagePath: String, maxWidthGet: Int, maxHeightGet: Int): Bitmap? {
+    // Scale image maintaining proportions
+    private fun scaleBitmap(bitmap: Bitmap, targetWidth: Int): Bitmap {
+        val scaleFactor = targetWidth.toFloat() / bitmap.width
+        val targetHeight = (bitmap.height * scaleFactor).toInt()
+        return Bitmap.createScaledBitmap(bitmap, targetWidth, targetHeight, true)
+    }
+
+    private fun compressImage(
+        imagePath: String,
+        maxWidthGet: Int,
+        maxHeightGet: Int,
+        keepAspectRatio: Boolean
+    ): Bitmap? {
 
         val maxHeight = maxWidthGet.toFloat()
         val maxWidth = maxHeightGet.toFloat()
@@ -93,7 +117,7 @@ class CreatePDFFromMultipleImage(getResult: MethodChannel.Result) {
         var imgRatio = actualWidth.toFloat() / actualHeight.toFloat()
         val maxRatio = maxWidth / maxHeight
 
-        if (actualHeight > maxHeight || actualWidth > maxWidth) {
+        if (keepAspectRatio && actualHeight > maxHeight) {
             if (imgRatio < maxRatio) {
                 imgRatio = maxHeight / actualHeight
                 actualWidth = (imgRatio * actualWidth).toInt()
@@ -118,8 +142,6 @@ class CreatePDFFromMultipleImage(getResult: MethodChannel.Result) {
 
         try {
             bmp = BitmapFactory.decodeFile(imagePath, options)
-            val baos = ByteArrayOutputStream()
-            bmp.compress(Bitmap.CompressFormat.JPEG, 10, baos)
         } catch (exception: OutOfMemoryError) {
             exception.printStackTrace()
             return null