+++ lib/models/pdf_from_multiple_image_config.dart
@@ -1,5 +1,3 @@
-import 'package:pdf_combiner/models/image_quality.dart';
-
 import 'image_scale.dart';
 
 /// Configuration for generating a PDF from multiple images.
@@ -13,7 +11,6 @@ class PdfFromMultipleImageConfig {
   /// Creates an instance of [PdfFromMultipleImageConfig].
   ///
   /// [rescale] allows specifying a scaling option for the images.
-  /// [compression] sets the quality level for the images, defaulting to [ImageQuality.high].
   /// [keepAspectRatio] determines if the aspect ratio should be preserved, defaulting to `true`.
   const PdfFromMultipleImageConfig({
     ImageScale? rescale,